Error fixes from models.Added email API

+++ src/main/java/com/techm/jobs/user/dao/CandidateRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.techm.jobs.user.model.Candidate;
 
-public interface CandidateRepository  extends JpaRepository<Candidate, Long>{
+public interface CandidateRepository  extends JpaRepository<Candidate, Integer>{
 
 }
